chore: Rename SwiftSearch to SwiftTypeSearch
