+++ packages/gatsby-theme-newrelic/src/styles/SwiftTypeSearchStyles.js
